# spec/platform/openapi.yaml
openapi: 3.0.3
...
info:
  title: dufflebag platform API
  version: "1.0.0"
  description: |
    Tenancy, service principals and project pins for a dufflebag registry.

    This is not an HCP-compatible surface and makes no attempt to be. The
    compatibility plane at /packer/2023-01-01/... and /resource-manager/... is
    where shape is dictated by clients we do not control.
  license:
    name: MPL-2.0
    url: https://www.mozilla.org/en-US/MPL/2.0/
servers:
  - url: /
...
security:
  - bearerAuth: []
paths:
...
  /api/v1/organizations:
    get:
      tags: [tenancy]
      operationId: listOrganizations
      summary: List organizations visible to the caller
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  organizations:
                    type: array
                    items: { $ref: "#/components/schemas/Organization" }
                required: [organizations]
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
    post:
      tags: [tenancy]
      operationId: createOrganization
      summary: Create an organization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string, minLength: 1, maxLength: 200 }
              required: [name]
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Organization" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409": { $ref: "#/components/responses/Conflict" }
...
  /api/v1/organizations/{organizationId}/projects:
    parameters:
      - $ref: "#/components/parameters/OrganizationId"
    get:
      tags: [tenancy]
      operationId: listProjects
      summary: List projects in an organization
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  projects:
                    type: array
                    items: { $ref: "#/components/schemas/Project" }
                required: [projects]
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
    post:
      tags: [tenancy]
      operationId: createProject
      summary: Create a project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string, minLength: 1, maxLength: 200 }
              required: [name]
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Project" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409": { $ref: "#/components/responses/Conflict" }
  /api/v1/organizations/{organizationId}/projects/{projectId}:
    parameters:
      - $ref: "#/components/parameters/OrganizationId"
      - $ref: "#/components/parameters/ProjectId"
    get:
      tags: [tenancy]
      operationId: getProject
      summary: Get a project
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Project" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
    delete:
      tags: [tenancy]
      operationId: deleteProject
      summary: Delete an empty project
      description: Refused while the project still contains buckets.
      responses:
        "204": { description: Deleted }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409":
          description: Still contains buckets.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        The same token the compatibility plane accepts. Everything downstream
        authorizes off verified claims rather than the credential that produced
        them, so a second issuer can be added without touching any consumer
        (ADR-0004).
  parameters:
    OrganizationId:
      name: organizationId
      in: path
      required: true
      description: |
        UUID. Organization and project identifiers appear in compatibility-plane
        paths and in HCP_ORGANIZATION_ID / HCP_PROJECT_ID, so they must look
        like what clients already expect — UUIDs, not the ULIDs used for
        registry resources.
      schema: { type: string, format: uuid }
    ProjectId:
      name: projectId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  responses:
    Unauthorized:
      description: Missing or invalid token.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Forbidden:
      description: |
        The caller is entitled to this tenancy but its role does not permit the
        operation. Distinct from 404, which is what a tenancy the caller may not
        see returns — confirming a tenancy exists is itself a disclosure, whereas
        a role refusal within a visible tenancy conceals nothing (ADR-0017).
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotFound:
      description: No such resource.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Conflict:
      description: Already exists.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
  schemas:
    Error:
      type: object
      description: |
        Deliberately NOT the compatibility plane's GoogleRPCStatus shape. This
        plane is ours and has no client regex-matching its error text.
      properties:
        message: { type: string }
        detail: { type: string }
      required: [message]
...
    Organization:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        created_at: { type: string, format: date-time }
      required: [id, name, created_at]
    Project:
      type: object
      properties:
        id: { type: string, format: uuid }
        organization_id: { type: string, format: uuid }
        name: { type: string }
        created_at:
          type: string
          format: date-time
          description: |
            Load-bearing. An unpinned Packer CLI selects the OLDEST project when
            several exist, so this timestamp determines default tenant selection
            and must never be rewritten (ADR-0003).
      required: [id, organization_id, name, created_at]
...